isolate AA drought specific legen items, setup empty legend for AA storm
