+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AAStormLegend/index.tsx
@@ -1,0 +1,11 @@
+import { Typography } from '@material-ui/core';
+
+function AAStormLegend() {
+  return (
+    <Typography>
+      AA storm Legend : nothing specific to AA storm to show yet
+    </Typography>
+  );
+}
+
+export default AAStormLegend;
